feat: add routes config for endpoints.
Make routes  folder with each endpoint configuration.
The routes are configured with blueprint of flask.
The prevous routes were moved to and config to the routes folder.

Reconfigure app.py with the routes change.

Add `tutor.py` in gitignore.
