+++ codejailservice/routes/__init__.py
@@ -1,0 +1,5 @@
+"""Init for `routes` to import all methods allowed for each route."""
+
+from .code_exec import *
+from .health import *
+from .index import *
