+++ codejailservice/routes/routes_config.py
@@ -1,0 +1,5 @@
+"""Configuration of blueprint in order to set  each route."""
+
+from flask import Blueprint
+
+routes = Blueprint('routes', __name__)
